restructure the directory
