+++ components/google-cloud/google_cloud_pipeline_components/aiplatform/model/upload_model/component.yaml
@@ -144,8 +144,8 @@ outputs:
 - {name: gcp_resources, type: String}
 implementation:
   container:
-    image: gcr.io/managed-pipeline-test/google-cloud-pipeline-components:v18
-    command: [python3, -u, -m, google_cloud_pipeline_components.experimental.remote.gcp_launcher.launcher]
+    image: gcr.io/ml-pipeline/google-cloud-pipeline-components:0.1.8
+    command: [python3, -u, -m, google_cloud_pipeline_components.remote.experimental.gcp_launcher.launcher]
     args: [
       --type, UploadModel,
       --payload,
