+++ components/google-cloud/google_cloud_pipeline_components/experimental/custom_job/component.yaml
@@ -22,7 +22,7 @@ outputs:
 implementation:
   container:
     image: gcr.io/ml-pipeline/google-cloud-pipeline-components:0.1.7
-    command: [python3, -u, -m, google_cloud_pipeline_components.experimental.remote.gcp_launcher.launcher]
+    command: [python3, -u, -m, google_cloud_pipeline_components.remote.experimental.gcp_launcher.launcher]
     args: [
       --type, CustomJob,
       --payload, {inputValue: payload},
