+++ components/google-cloud/google_cloud_pipeline_components/experimental/custom_job/custom_job.py
@@ -256,7 +256,7 @@ def custom_training_job_op(
                 image=_DEFAULT_CUSTOM_JOB_CONTAINER_IMAGE,
                 command=[
                     "python3", "-u", "-m",
-                    "google_cloud_pipeline_components.experimental.remote.gcp_launcher.launcher"
+                    "google_cloud_pipeline_components.remote.experimental.gcp_launcher.launcher"
                 ],
                 args=[
                     '--type',
