+++ components/google-cloud/google_cloud_pipeline_components/experimental/wait_gcp_resources/component.yaml
@@ -46,7 +46,7 @@ outputs:
 implementation:
   container:
     image: gcr.io/ml-pipeline/google-cloud-pipeline-components:latest
-    command: [python3, -u, -m, google_cloud_pipeline_components.experimental.remote.gcp_launcher.launcher]
+    command: [python3, -u, -m, google_cloud_pipeline_components.remote.experimental.gcp_launcher.launcher]
     args: [
       --type, Wait,
       # project and location are already specified in the input gcp_resources
