+++ components/google-cloud/google_cloud_pipeline_components/remote/experimental/gcp_launcher/utils/artifact_util.py
