+++ components/google-cloud/tests/aiplatform/testdata/model_upload_pipeline.json
@@ -100,9 +100,9 @@
                     "python3",
                     "-u",
                     "-m",
-                    "google_cloud_pipeline_components.experimental.remote.gcp_launcher.launcher"
+                    "google_cloud_pipeline_components.remote.experimental.gcp_launcher.launcher"
                  ],
-                 "image":"gcr.io/managed-pipeline-test/google-cloud-pipeline-components:v18"
+                 "image":"gcr.io/ml-pipeline/google-cloud-pipeline-components:0.1.8"
               }
            }
         }
