+++ components/google-cloud/tests/experimental/custom_job/test_custom_job.py
@@ -105,7 +105,7 @@ implementation:
                         'test_launcher_image',
                     'command': [
                         'python3', '-u', '-m',
-                        'google_cloud_pipeline_components.experimental.remote.gcp_launcher.launcher'
+                        'google_cloud_pipeline_components.remote.experimental.gcp_launcher.launcher'
                     ],
                     'args': [
                         '--type', 'CustomJob', '--payload',
@@ -159,7 +159,7 @@ implementation:
                         'gcr.io/tfe-ecosystem-dev/temp-custom-job:latest',
                     'command': [
                         'python3', '-u', '-m',
-                        'google_cloud_pipeline_components.experimental.remote.gcp_launcher.launcher'
+                        'google_cloud_pipeline_components.remote.experimental.gcp_launcher.launcher'
                     ],
                     'args': [
                         '--type', 'CustomJob', '--project', {
@@ -202,7 +202,7 @@ implementation:
                         'test_launcher_image',
                     'command': [
                         'python3', '-u', '-m',
-                        'google_cloud_pipeline_components.experimental.remote.gcp_launcher.launcher'
+                        'google_cloud_pipeline_components.remote.experimental.gcp_launcher.launcher'
                     ],
                     'args': [
                         '--type', 'CustomJob', '--payload',
@@ -237,7 +237,7 @@ implementation:
                         'test_launcher_image',
                     'command': [
                         'python3', '-u', '-m',
-                        'google_cloud_pipeline_components.experimental.remote.gcp_launcher.launcher'
+                        'google_cloud_pipeline_components.remote.experimental.gcp_launcher.launcher'
                     ],
                     'args': [
                         '--type', 'CustomJob', '--payload',
@@ -271,7 +271,7 @@ implementation:
                         'test_launcher_image',
                     'command': [
                         'python3', '-u', '-m',
-                        'google_cloud_pipeline_components.experimental.remote.gcp_launcher.launcher'
+                        'google_cloud_pipeline_components.remote.experimental.gcp_launcher.launcher'
                     ],
                     'args': [
                         '--type', 'CustomJob', '--payload',
@@ -307,7 +307,7 @@ implementation:
                         'test_launcher_image',
                     'command': [
                         'python3', '-u', '-m',
-                        'google_cloud_pipeline_components.experimental.remote.gcp_launcher.launcher'
+                        'google_cloud_pipeline_components.remote.experimental.gcp_launcher.launcher'
                     ],
                     'args': [
                         '--type', 'CustomJob', '--payload',
@@ -341,7 +341,7 @@ implementation:
                         'test_launcher_image',
                     'command': [
                         'python3', '-u', '-m',
-                        'google_cloud_pipeline_components.experimental.remote.gcp_launcher.launcher'
+                        'google_cloud_pipeline_components.remote.experimental.gcp_launcher.launcher'
                     ],
                     'args': [
                         '--type', 'CustomJob', '--payload',
@@ -374,7 +374,7 @@ implementation:
                         'test_launcher_image',
                     'command': [
                         'python3', '-u', '-m',
-                        'google_cloud_pipeline_components.experimental.remote.gcp_launcher.launcher'
+                        'google_cloud_pipeline_components.remote.experimental.gcp_launcher.launcher'
                     ],
                     'args': [
                         '--type', 'CustomJob', '--payload',
@@ -408,7 +408,7 @@ implementation:
                         'test_launcher_image',
                     'command': [
                         'python3', '-u', '-m',
-                        'google_cloud_pipeline_components.experimental.remote.gcp_launcher.launcher'
+                        'google_cloud_pipeline_components.remote.experimental.gcp_launcher.launcher'
                     ],
                     'args': [
                         '--type', 'CustomJob', '--payload',
@@ -442,7 +442,7 @@ implementation:
                         'test_launcher_image',
                     'command': [
                         'python3', '-u', '-m',
-                        'google_cloud_pipeline_components.experimental.remote.gcp_launcher.launcher'
+                        'google_cloud_pipeline_components.remote.experimental.gcp_launcher.launcher'
                     ],
                     'args': [
                         '--type', 'CustomJob', '--payload',
@@ -475,7 +475,7 @@ implementation:
                         'test_launcher_image',
                     'command': [
                         'python3', '-u', '-m',
-                        'google_cloud_pipeline_components.experimental.remote.gcp_launcher.launcher'
+                        'google_cloud_pipeline_components.remote.experimental.gcp_launcher.launcher'
                     ],
                     'args': [
                         '--type', 'CustomJob', '--payload',
@@ -522,7 +522,7 @@ implementation:
                         'test_launcher_image',
                     'command': [
                         'python3', '-u', '-m',
-                        'google_cloud_pipeline_components.experimental.remote.gcp_launcher.launcher'
+                        'google_cloud_pipeline_components.remote.experimental.gcp_launcher.launcher'
                     ],
                     'args': [
                         '--type', 'CustomJob', '--payload',
@@ -555,7 +555,7 @@ implementation:
                         'test_launcher_image',
                     'command': [
                         'python3', '-u', '-m',
-                        'google_cloud_pipeline_components.experimental.remote.gcp_launcher.launcher'
+                        'google_cloud_pipeline_components.remote.experimental.gcp_launcher.launcher'
                     ],
                     'args': [
                         '--type', 'CustomJob', '--payload',
