+++ components/google-cloud/tests/remote/experimental/gcp_launcher/utils/test_artifact_util.py
@@ -17,12 +17,12 @@ import os
 import time
 import unittest
 from unittest import mock
-from google_cloud_pipeline_components.experimental.remote.gcp_launcher import custom_job_remote_runner
+from google_cloud_pipeline_components.remote.experimental.gcp_launcher import custom_job_remote_runner
 from google.cloud import aiplatform
 from google.cloud.aiplatform.compat.types import job_state as gca_job_state
 from google_cloud_pipeline_components.proto.gcp_resources_pb2 import GcpResources
 from google.protobuf import json_format
-from google_cloud_pipeline_components.experimental.remote.gcp_launcher.utils import artifact_util
+from google_cloud_pipeline_components.remote.experimental.gcp_launcher.utils import artifact_util
 
 
 class ArtifactUtilTests(unittest.TestCase):
