+++ components/google-cloud/tests/remote/experimental/gcp_launcher/utils/test_json_util.py
@@ -16,7 +16,7 @@
 import json
 import unittest
 from unittest import mock
-from google_cloud_pipeline_components.experimental.remote.gcp_launcher.utils import json_util
+from google_cloud_pipeline_components.remote.experimental.gcp_launcher.utils import json_util
 
 
 class JsonUtilTests(unittest.TestCase):
